MBS-11392 move StubDevice to `test-runner:service` module
