+++ subprojects/settings.gradle.kts
@@ -33,7 +33,6 @@ include(":gradle:impact")
 include(":test-runner:client")
 include(":test-runner:device-provider")
 include(":test-runner:report")
-include(":test-runner:stub")
 include(":test-runner:service")
 include(":test-runner:command-line-executor")
 include(":gradle:sentry-config")
