+++ subprojects/test-runner/client/src/testFixtures/kotlin/com/avito/runner/config/StubInstrumentationTestsActionParams.kt
@@ -2,8 +2,8 @@ package com.avito.runner.config
 
 import com.avito.android.stats.StatsDConfig
 import com.avito.logger.LoggerFactory
-import com.avito.runner.scheduler.runner.model.ExecutionParameters
 import com.avito.runner.scheduler.runner.createStubInstance
+import com.avito.runner.scheduler.runner.model.ExecutionParameters
 import com.avito.runner.scheduler.suite.filter.ImpactAnalysisResult
 import com.avito.runner.scheduler.suite.filter.createStubInstance
 import com.avito.utils.gradle.KubernetesCredentials
