+++ subprojects/test-runner/service/src/main/kotlin/com/avito/runner/service/worker/device/stub/StubActionResult.kt
@@ -1,4 +1,4 @@
-package com.avito.runner.service.worker.device
+package com.avito.runner.service.worker.device.stub
 
 sealed class StubActionResult<T> {
