+++ subprojects/test-runner/service/src/test/kotlin/com/avito/runner/service/DeviceWorkerPoolTest.kt
@@ -12,13 +12,12 @@ import com.avito.runner.service.model.intention.createStubInstance
 import com.avito.runner.service.worker.device.Device
 import com.avito.runner.service.worker.device.Device.DeviceStatus
 import com.avito.runner.service.worker.device.DeviceCoordinate
-import com.avito.runner.service.worker.device.StubActionResult
-import com.avito.runner.service.worker.device.StubDevice
-import com.avito.runner.service.worker.device.StubDevice.Companion.installApplicationSuccess
 import com.avito.runner.service.worker.device.createStubInstance
+import com.avito.runner.service.worker.device.stub.StubActionResult
+import com.avito.runner.service.worker.device.stub.StubDevice
+import com.avito.runner.service.worker.device.stub.StubDevice.Companion.installApplicationSuccess
 import com.avito.runner.service.worker.listener.StubDeviceListener
 import com.avito.test.TestDispatcher
-import com.avito.test.listWithDefault
 import com.avito.test.receiveAvailable
 import com.avito.time.StubTimeProvider
 import com.google.common.truth.Truth.assertWithMessage
@@ -78,10 +77,7 @@ class DeviceWorkerPoolTest {
                     installApplicationSuccess(), // Install application
                     installApplicationSuccess() // Install test application
                 ),
-                gettingDeviceStatusResults = listWithDefault(
-                    1 + intentions.size,
-                    DeviceStatus.Alive
-                ),
+                gettingDeviceStatusResults = List(1 + intentions.size) { DeviceStatus.Alive },
                 clearPackageResults = (0 until intentions.size - 1).flatMap {
                     listOf(
                         StubActionResult.Success(Result.Success(Unit)),
