--- subprojects/test-runner/service/src/test/kotlin/com/avito/test/ListWithDefault.kt
@@ -1,5 +1,0 @@
-package com.avito.test
-
-internal fun <T : Any> listWithDefault(size: Int, value: T): List<T> {
-    return (0 until size).map { value }
-}
